+++ tests/engine/test_flow_runner.py
@@ -20,9 +20,11 @@ from prefect.engine.state import (
     CachedState,
     Failed,
     Finished,
-    Pending,
-    Retrying,
     Mapped,
+    Paused,
+    Pending,
+    Resume,
+    Retrying,
     Running,
     Scheduled,
     Skipped,
@@ -500,7 +502,6 @@ class TestRunFlowStep:
             state=Running(),
             task_states={},
             start_tasks=[],
-            start_task_ids=[],
             return_tasks=set(),
             task_runner_state_handlers=[],
             executor=LocalExecutor(),
@@ -517,7 +518,6 @@ class TestRunFlowStep:
                 state=state,
                 task_states={},
                 start_tasks=[],
-                start_task_ids=[],
                 return_tasks=set(),
                 task_runner_state_handlers=[],
                 executor=Executor(),
@@ -533,7 +533,6 @@ class TestRunFlowStep:
             state=Running(),
             task_states={},
             start_tasks=[],
-            start_task_ids=[],
             return_tasks=set(),
             task_runner_state_handlers=[],
             executor=LocalExecutor(),
@@ -554,45 +553,6 @@ class TestStartTasks:
         assert "Task not evaluated" in state.result[t2].message
         assert "Task not evaluated" not in state.result[t3].message
 
-    def test_start_tasks_ids_are_respected(self):
-        f = Flow()
-        t1, t2, t3 = Task(), Task(), Task()
-        f.add_edge(t1, t2)
-        f.add_edge(t2, t3)
-        state = FlowRunner(flow=f).run(
-            start_task_ids=[t3.id], return_tasks=[t1, t2, t3]
-        )
-
-        assert "Task not evaluated" in state.result[t1].message
-        assert "Task not evaluated" in state.result[t2].message
-        assert "Task not evaluated" not in state.result[t3].message
-
-    def test_start_tasks_ids_can_be_combined_with_start_tasks(self):
-        f = Flow()
-        t1, t2, t3 = Task(), Task(), Task()
-        f.add_edge(t1, t2)
-        f.add_edge(t2, t3)
-        state = FlowRunner(flow=f).run(
-            start_tasks=[t2], start_task_ids=[t3.id], return_tasks=[t1, t2, t3]
-        )
-
-        assert "Task not evaluated" in state.result[t1].message
-        assert "Task not evaluated" not in state.result[t2].message
-        assert "Task not evaluated" not in state.result[t3].message
-
-    def test_start_tasks_ids_can_overlap_with_start_tasks(self):
-        f = Flow()
-        t1, t2, t3 = Task(), Task(), Task()
-        f.add_edge(t1, t2)
-        f.add_edge(t2, t3)
-        state = FlowRunner(flow=f).run(
-            start_tasks=[t2], start_task_ids=[t2.id], return_tasks=[t1, t2, t3]
-        )
-
-        assert "Task not evaluated" in state.result[t1].message
-        assert "Task not evaluated" not in state.result[t2].message
-        assert "Task not evaluated" not in state.result[t3].message
-
     def test_invalid_start_task(self):
         f = Flow()
         t1 = Task()
@@ -602,16 +562,6 @@ class TestStartTasks:
 
         assert state.is_failed()
         assert "not found in Flow" in str(state.result)
-
-    def test_invalid_start_task_id(self):
-        f = Flow()
-        t1 = Task()
-        f.add_task(t1)
-
-        state = FlowRunner(flow=f).run(start_task_ids=["nope"])
-
-        assert state.is_failed()
-        assert "Invalid start_task_ids" in str(state.result)
 
     def test_start_tasks_ignores_triggers(self):
         f = Flow()
@@ -732,7 +682,9 @@ class TestInputCaching:
             parameters=dict(x=1),
             return_tasks=[res],
             start_tasks=[res],
-            task_states={res: first_state.result[res]},
+            task_states={
+                res: Resume(cached_inputs=first_state.result[res].cached_inputs)
+            },
         )
         assert isinstance(second_state, Success)
         assert second_state.result[res].result == 12
@@ -1354,3 +1306,12 @@ class TestMapping:
         )
         assert state.is_running()
         assert isinstance(state.result[res], Scheduled)
+
+
+def test_paused_tasks_stay_paused_when_run():
+    t = Task()
+    f = Flow(tasks=[t])
+
+    state = f.run(task_states={t: Paused()}, return_tasks=[t])
+    assert state.is_running()
+    assert isinstance(state.result[t], Paused)